Move TestCluster plugin and module setup from Node to Cluster
