+++ build-tools/src/main/java/org/elasticsearch/gradle/testclusters/ElasticsearchNode.java
@@ -21,18 +21,13 @@ import org.elasticsearch.gradle.ReaperService;
 import org.elasticsearch.gradle.Version;
 import org.elasticsearch.gradle.VersionProperties;
 import org.elasticsearch.gradle.distribution.ElasticsearchDistributionTypes;
-import org.elasticsearch.gradle.transform.UnzipTransform;
 import org.elasticsearch.gradle.util.Pair;
 import org.gradle.api.Action;
 import org.gradle.api.Named;
 import org.gradle.api.NamedDomainObjectContainer;
 import org.gradle.api.Project;
 import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.Dependency;
-import org.gradle.api.artifacts.type.ArtifactTypeDefinition;
-import org.gradle.api.attributes.Attribute;
 import org.gradle.api.file.ArchiveOperations;
-import org.gradle.api.file.ConfigurableFileCollection;
 import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.FileSystemOperations;
 import org.gradle.api.file.FileTree;
@@ -98,7 +93,6 @@ import java.util.stream.Stream;
 
 import static java.util.Objects.requireNonNull;
 import static java.util.Optional.ofNullable;
-import static org.elasticsearch.gradle.plugin.BasePluginBuildPlugin.EXPLODED_BUNDLE_CONFIG;
 
 public class ElasticsearchNode implements TestClusterConfiguration {
 
@@ -141,7 +135,6 @@ public class ElasticsearchNode implements TestClusterConfiguration {
 
     private final LinkedHashMap<String, Predicate<TestClusterConfiguration>> waitConditions = new LinkedHashMap<>();
     private final Map<String, Configuration> pluginAndModuleConfigurations = new HashMap<>();
-    private final ConfigurableFileCollection pluginAndModuleConfiguration;
     private final List<Provider<File>> plugins = new ArrayList<>();
     private final List<Provider<File>> modules = new ArrayList<>();
     private final LazyPropertyMap<String, CharSequence> settings = new LazyPropertyMap<>("Settings", this);
@@ -171,8 +164,6 @@ public class ElasticsearchNode implements TestClusterConfiguration {
     private final Provider<File> runtimeJava;
     private final Function<Version, Boolean> isReleasedVersion;
     private final List<ElasticsearchDistribution> distributions = new ArrayList<>();
-    private final Attribute<Boolean> bundleAttribute = Attribute.of("bundle", Boolean.class);
-
     private int currentDistro = 0;
     private TestDistribution testDistribution;
     private volatile Process esProcess;
@@ -223,10 +214,8 @@ public class ElasticsearchNode implements TestClusterConfiguration {
         waitConditions.put("ports files", this::checkPortsFilesExistWithDelay);
         defaultConfig.put("cluster.name", clusterName);
 
-        pluginAndModuleConfiguration = project.getObjects().fileCollection();
         setTestDistribution(TestDistribution.INTEG_TEST);
         setVersion(VersionProperties.getElasticsearch());
-        configureArtifactTransforms();
     }
 
     @Input
@@ -308,36 +297,16 @@ public class ElasticsearchNode implements TestClusterConfiguration {
         return pluginAndModuleConfigurations.values();
     }
 
-    // creates a configuration to depend on the given plugin project, then wraps that configuration
-    // to grab the zip as a file provider
-    private Provider<RegularFile> maybeCreatePluginOrModuleDependency(String path, String consumingConfiguration) {
-        var configuration = pluginAndModuleConfigurations.computeIfAbsent(path, key -> {
-            var bundleDependency = this.project.getDependencies().project(Map.of("path", path, "configuration", consumingConfiguration));
-            return project.getConfigurations().detachedConfiguration(bundleDependency);
-        });
-
-        Provider<File> fileProvider = configuration.getElements()
-            .map(
-                s -> s.stream()
-                    .findFirst()
-                    .orElseThrow(
-                        () -> new IllegalStateException(consumingConfiguration + " configuration of project " + path + " had no files")
-                    )
-                    .getAsFile()
-            );
-        return project.getLayout().file(fileProvider);
-    }
-
     @Override
     public void plugin(Provider<RegularFile> plugin) {
         checkFrozen();
-        registerExtractedConfig(plugin);
+        //
         this.plugins.add(plugin.map(RegularFile::getAsFile));
     }
 
     @Override
     public void plugin(String pluginProjectPath) {
-        plugin(maybeCreatePluginOrModuleDependency(pluginProjectPath, "zip"));
+        throw new IllegalStateException("Deprecated");
     }
 
     public void plugin(TaskProvider<Zip> plugin) {
@@ -347,7 +316,7 @@ public class ElasticsearchNode implements TestClusterConfiguration {
     @Override
     public void module(Provider<RegularFile> module) {
         checkFrozen();
-        registerExtractedConfig(module);
+        //registerExtractedConfig(module);
         this.modules.add(module.map(RegularFile::getAsFile));
     }
 
@@ -355,31 +324,9 @@ public class ElasticsearchNode implements TestClusterConfiguration {
         module(project.getLayout().file(module.map(Sync::getDestinationDir)));
     }
 
-    private void registerExtractedConfig(Provider<RegularFile> pluginProvider) {
-        Dependency pluginDependency = this.project.getDependencies().create(project.files(pluginProvider));
-        Configuration extractedConfig = project.getConfigurations().detachedConfiguration(pluginDependency);
-        extractedConfig.getAttributes().attribute(ArtifactTypeDefinition.ARTIFACT_TYPE_ATTRIBUTE, ArtifactTypeDefinition.DIRECTORY_TYPE);
-        extractedConfig.getAttributes().attribute(bundleAttribute, true);
-        pluginAndModuleConfiguration.from(extractedConfig);
-    }
-
-    private void configureArtifactTransforms() {
-        project.getDependencies().getAttributesSchema().attribute(bundleAttribute);
-        project.getDependencies().getArtifactTypes().maybeCreate(ArtifactTypeDefinition.ZIP_TYPE);
-        project.getDependencies().registerTransform(UnzipTransform.class, transformSpec -> {
-            transformSpec.getFrom()
-                .attribute(ArtifactTypeDefinition.ARTIFACT_TYPE_ATTRIBUTE, ArtifactTypeDefinition.ZIP_TYPE)
-                .attribute(bundleAttribute, true);
-            transformSpec.getTo()
-                .attribute(ArtifactTypeDefinition.ARTIFACT_TYPE_ATTRIBUTE, ArtifactTypeDefinition.DIRECTORY_TYPE)
-                .attribute(bundleAttribute, true);
-            transformSpec.getParameters().setAsFiletreeOutput(true);
-        });
-    }
-
     @Override
     public void module(String moduleProjectPath) {
-        module(maybeCreatePluginOrModuleDependency(moduleProjectPath, EXPLODED_BUNDLE_CONFIG));
+        throw new IllegalStateException("Deprecated");
     }
 
     @Override
@@ -1549,17 +1496,6 @@ public class ElasticsearchNode implements TestClusterConfiguration {
 
     private Path getExtractedDistributionDir() {
         return distributions.get(currentDistro).getExtracted().getSingleFile().toPath();
-    }
-
-    @Classpath
-    public FileCollection getInstalledClasspath() {
-        return pluginAndModuleConfiguration.getAsFileTree().filter(f -> f.getName().endsWith(".jar"));
-    }
-
-    @InputFiles
-    @PathSensitive(PathSensitivity.RELATIVE)
-    public FileCollection getInstalledFiles() {
-        return pluginAndModuleConfiguration.getAsFileTree().filter(f -> f.getName().endsWith(".jar") == false);
     }
 
     @Classpath
